Move ButtonContentSupplier to menu package
